feat: simplified async example
